Update test and ficture files - 100% cov

+++ spec/requests/api/v1/forecast_index_spec.rb
@@ -60,7 +60,7 @@ RSpec.describe 'Forecast API - Endpoints', type: :request do
     expect(day).to have_key :sunset
     expect(day[:sunset]).to be_a String
     expect(day).to have_key :max_temp
-    expect(day[:max_temp]).to be_a Integer
+    expect(day[:max_temp]).to be_a Float
     expect(day).to have_key :min_temp
     expect(day[:min_temp]).to be_a Float
     expect(day).to have_key :conditions
